Handle RFC2047 words split in the middle of a multibyte character
This is the first of a series of commits that I plan to do to refactor
and simplify RFC2047 code.  This first commit fixes the issue reported
in #1015, but is still lacking some of the features and corner-case
checks in the old code. Especially, support for IgnoreLinearWhiteSpace
is has not been added back yet.

Issue #1015